rkt: Move image related functions to image.go
